requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
bson==0.5.8
dataclasses==0.6
Django==3.2.8
django-cors-headers==3.10.1
django-prometheus==2.2.0
djangorestframework==3.12.4
djongo==1.3.6
drf-nested-routers==0.93.4
prometheus-client==0.12.0
psycopg2-binary==2.8.6
pymongo==3.11.4
PyOpenGL==3.1.5
PySide2==5.15.2
python-dateutil==2.8.2
pytz==2021.3
shiboken2==5.15.2
six==1.16.0
sqlparse==0.2.4
tzdata==2021.5
drf_yasg==1.20.0
